requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.7.1
django==5.1.7
django-prometheus==2.3.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
iniconfig==2.1.0
kombu==5.5.0
packaging==24.2
pluggy==1.5.0
prometheus-client==0.21.1
promise==2.3
prompt-toolkit==3.0.50
psycopg2-binary==2.9.10
pyjwt==2.9.0
pytest==8.3.5
pytest-cov==6.0.0
pytest-django==4.10.0
python-dateutil==2.9.0.post0
pyyaml==6.0.2
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing-extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
